# shared-schemas/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Central account and token service for the Ducktivity suite: passwordless email-OTP login and the SOLE issuer of session tokens. The login endpoints below are the only client-facing contract other apps consume (cross-origin, pre-auth); every token-authenticated endpoint lives on an app's own backend, which merely verifies the token minted here.",
        "title": "Ducktivity Identity API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/auth/request-code": {
            "post": {
                "description": "Emails a 6-digit login code, creating the account if the address is new. Always returns a generic acknowledgement so it cannot reveal whether an account exists.",
                "tags": [
                    "auth"
                ],
                "summary": "Request a login code",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.AuthRequestCodeInput"
                            }
                        }
                    },
                    "description": "Email to send the login code to",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Deliberately vague acknowledgement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.MessageResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Asked again before the resend cooldown elapsed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/verify-code": {
            "post": {
                "description": "Exchanges an email + 6-digit code for a session token. The token carries the account's current suite-wide entitlement, so every app learns paid access from the token alone.",
                "tags": [
                    "auth"
                ],
                "summary": "Verify a login code",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.AuthVerifyCodeInput"
                            }
                        }
                    },
                    "description": "Email and the code that was emailed",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "The session token plus the account it belongs to",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.VerifyResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Malformed body or invalid email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
        "schemas": {
            "api.AuthRequestCodeInput": {
                "type": "object",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    }
                }
            },
            "api.AuthVerifyCodeInput": {
                "type": "object",
                "required": [
                    "code",
                    "email"
                ],
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
            "api.ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
            "api.MessageResponse": {
                "type": "object",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "api.User": {
                "type": "object",
                "required": [
                    "email",
                    "id"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    }
                }
            },
            "api.VerifyResponse": {
                "type": "object",
                "required": [
                    "token",
                    "user"
                ],
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/api.User"
                    }
                }
            }
        }
    }
}